Initial checkin of new USJobs plugin
-- adds new method to JobsSitePlugin called parseJobsForPageFromFile.
This method takes a file name and returns the array of jobs parsed from
the file.  It contains a subset of the code that used to be in the
_getMyJobsFromHTMLFiles_ method (which now calls
parseJobsForPageFromFile.)   This split allows plugins to override how
they parse a particular local file.  In this case, it allows USJobs to
call a Microdata parser method on the file to get the listings.

-- Also added some attempted improvements to logging for osascript
calls.